DbPage: Fix Qt regression on menu buttons
Due to some regression in Qt 6.6, QPushButtons with a QMenu attached to
them have a visual glitch making them look like they are stuck in the
pressed state. To overcome this issue, I manually implemented menu
buttons. This change caused the buttons to lose their tiny triangles at
the end indicating they are menu buttons, but who cares? Not me.