styles: Force normal font-weight for buttons
For some reason, Qt uses the GTK3 theme settings for the unset styles.
This is a big problem, especially for the users of the adw-gtk3 theme
which prefers a very bold font-weight for the buttons, rendering all qt
buttons extremely ugly. After countless hours of debugging, I was able
to triage and fix this issue with this simple solution.